requirements.txt
django-cms<3.3
django-treebeard>=4.0

https://github.com/divio/djangocms-column/archive/master.zip?1470356950.34
https://github.com/divio/djangocms-googlemap/archive/master.zip?1470356950.34
https://github.com/divio/djangocms-inherit/archive/master.zip?1470356950.34
https://github.com/divio/djangocms-link/archive/master.zip?1470356950.34
https://github.com/divio/djangocms-style/archive/master.zip?1470356950.34
easy_thumbnails
django-filer>=1.2
cmsplugin-filer>=1.1
djangocms-admin-style>=1.0.6
djangocms-text-ckeditor>=2.8.1,<3.0
Django<1.9
pytz
django-reversion>=1.8.7,<1.9
django-mptt<0.9
django-classy-tags>=0.3.4.1
html5lib>=0.90,!=0.9999,!=0.99999,<0.99999999
Pillow>=2.3
django-sekizai>=0.7
six
django-photologue>=3.5.1
bpython>=0.12
django-fontawesome==0.2.6
django-extensions==1.6.7